Implement edit project functionality
Edit project button from within project display will bring up the same modal but
with slightly different logic for the submit button. The logic is changed by setUpModal()
in the controller, which is called by the painter when either all projects or a single
project is displayed. The logic for the delete and close buttons is set once during
initialize().